use alerts, rather than toasts in FriendlyHttpResponseHandler
If notifications are disabled for an application, toasts are also
squelched, leaving important error messages on the floor.

One possible thing to look out for is that you currently pass a single
Context to the handler.  Previously, you could use the application's
context.  This meant if you got the async response when viewing another
activity, it could still be handled as the application Context still
existed.  Now, you have to use an activity's context, which means if the
response returns when viewing another activity, you'll get an exception.

A workaround would be to use 2 contexts, one for the application to
handle the response at all costs, and another for the activity to
display the error when possible (if the activity still exists).  We'll
see if that's necessary...